Rename footer to .jsx
